% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{ReproducibilitySodhi2023-package}
\alias{ReproducibilitySodhi2023}
\alias{ReproducibilitySodhi2023-package}
\title{ReproducibilitySodhi2023: Reproduction of Sodhi Study in JAMA 2023}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
